@@ -21,20 +21,21 @@
     "@mui/styled-engine-sc": "^6.0.0-alpha.5",
     "@mui/system": "^5.8.0",
     "@mui/x-date-pickers": "^6.18.1",
-    "@reduxjs/toolkit": "^1.9.5",
+    "@reduxjs/toolkit": "^2.2.1",
     "@svgr/webpack": "^8.0.1",
+    "@types/estree": "^1.0.5",
     "axios": "^1.6.2",
     "graphql": "^16.6.0",
     "moment": "^2.29.4",
-    "next": "13.4.4",
+    "next": "^14.1.0",
     "next-redux-wrapper": "^8.1.0",
-    "react": "18.2.0",
+    "react": "^18.2.0",
     "react-csv": "^2.2.2",
-    "react-dom": "18.2.0",
+    "react-dom": "^18.3.0-next-fecc288b7-20221025",
     "react-dropzone": "^14.2.3",
     "react-pdf": "^7.1.1",
-    "react-redux": "^8.0.7",
-    "react-toastify": "^9.1.3",
+    "react-redux": "^9.1.0",
+    "react-toastify": "^10.0.4",
     "react-use-websocket": "^4.5.0",
     "styled-components": "^6.1.1",
     "utif": "^3.1.0",
@@ -42,15 +43,14 @@
     "webpack": "^5.86.0"
   },
   "devDependencies": {
-    "@types/node": "20.2.6",
-    "@types/react": "18.2.11",
+    "@types/node": "20.11.20",
+    "@types/react": "18.2.58",
     "@types/react-csv": "^1.1.3",
-    "@types/react-dom": "18.2.4",
-    "@types/react-pdf": "^6.2.0",
+    "@types/react-dom": "18.2.19",
     "@types/utif": "^3.0.1",
     "@types/uuid": "^9.0.2",
-    "eslint": "8.42.0",
-    "eslint-config-next": "13.4.4",
-    "typescript": "5.1.3"
+    "eslint": "8.57.0",
+    "eslint-config-next": "14.1.0",
+    "typescript": "5.3.3"
   }
 }
